--- molecules/banner/data.ts
@@ -1,26 +1,0 @@
-import { Hashtag } from '~/@types/Hashtag';
-
-const mockData = [
-  {
-    title: '비사이드',
-    url: '#',
-  },
-  {
-    title: 'IT위키피디아테스트문구',
-    url: '##',
-  },
-  {
-    title: 'Share',
-    url: '###',
-  },
-  {
-    title: 'Test',
-    url: '####',
-  },
-  {
-    title: '니트 가이드',
-    url: '#####',
-  },
-] as ReadonlyArray<Hashtag>;
-
-export default mockData;
