--- molecules/banner/index.ts
@@ -1,2 +1,0 @@
-export { default as SearchBanner } from './SearchBanner';
-export { default as GuideBanner } from './GuideBanner';
