+++ molecules/main/BannerSection.tsx
@@ -1,5 +1,5 @@
 import { Section } from '~/atoms/layout';
-import { GuideBanner } from '~/molecules/banner';
+import GuideBanner from './GuideBanner';
 
 export default function BannerSection() {
   return (
