+++ molecules/main/FeaturedSection.module.scss
