+++ molecules/main/FeaturedSection.tsx
