+++ molecules/main/GuideBanner.styled.ts
