+++ molecules/main/RecentlyModifiedItem.styled.ts
