+++ molecules/main/SearchBanner.module.scss
