+++ molecules/main/index.ts
@@ -1,3 +1,4 @@
+export { default as SearchBanner } from './SearchBanner';
 export { default as MostViewSection } from './MostViewSection';
 export { default as FeaturedSection } from './FeaturedSection';
 export { default as BannerSection } from './BannerSection';
